isolate browser environment stubbing
